backend/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
six==1.16.0
uWSGI==2.0.21
Werkzeug==2.2.2
